+++ examples/Cargo.toml
@@ -1,0 +1,5 @@
+[workspace]
+
+members = ["*"]
+exclude = ["target"]
+resolver = "2"
