+++ examples/value_marker/src/main.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::{App, Commands, Component, Query};
 use bevy_ecs_markers::{
     params::{Marker, MarkerMut},
-    EntityMarker,
+    EntityMarker, InitMarkerData,
 };
 
 #[derive(EntityMarker)]
@@ -10,44 +10,27 @@ enum Players {
     Blue,
 }
 
-#[derive(EntityMarker)]
-struct CurrentPlayer;
-
 #[derive(Component)]
 struct Player(u32);
 
 fn main() {
     App::new()
-        .init_resource::<<Players as EntityMarker>::MarkerData>()
-        .init_resource::<<CurrentPlayer as EntityMarker>::MarkerData>()
+        .init_markerdata::<Players>()
         .add_startup_system(setup)
         .add_system(get_red_player)
-        .add_system(get_current_player)
         .run();
 }
 
-fn setup(
-    mut commands: Commands,
-    mut markers: MarkerMut<Players>,
-    mut current: MarkerMut<CurrentPlayer>,
-) {
+fn setup(mut commands: Commands, mut markers: MarkerMut<Players>) {
     let red = commands.spawn(Player(12)).id();
     markers[Players::Red] = red;
 
     let blue = commands.spawn(Player(7)).id();
     markers[Players::Blue] = blue;
-
-    *current = blue;
 }
 
 fn get_red_player(mut query: Query<&mut Player>, markers: Marker<Players>) {
     if let Ok(mut player) = query.get_mut(markers[Players::Red]) {
         player.0 = 15;
-    }
-}
-
-fn get_current_player(mut query: Query<&mut Player>, current: Marker<CurrentPlayer>) {
-    if let Ok(mut player) = query.get_mut(*current) {
-        player.0 = 2;
     }
 }
